Correlation functions in a new flexible format
Next job is to fix SpectralDensity in the same manner. We do this here,
because we need to fix saving of CorrelationFunctions in order to save
Molecules and Aggregates. To make them Saveable will require to
implement saving of Lists/Tuples and Dictionaries. Here we investigate
how flexible than saving has to be.